requirements.txt
Flask==2.0.1
openai==0.27.0
atlassian-python-api==3.41.1
PyGithub==1.58.0
python-dotenv==1.0.0
requests==2.31.0
openapi-spec-validator==0.6.0
fastapi==0.95.0
uvicorn==0.22.0
pydantic==2.0.0
